[Console] fix input stream related tests
