+++ src/Symfony/Component/Console/Tests/Helper/AbstractQuestionHelperTest.php
@@ -1,0 +1,33 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Console\Tests\Helper;
+
+use Symfony\Component\Console\Input\StreamableInputInterface;
+
+abstract class AbstractQuestionHelperTest extends \PHPUnit_Framework_TestCase
+{
+    protected function createStreamableInputInterfaceMock($stream = null, $interactive = true)
+    {
+        $mock = $this->getMock(StreamableInputInterface::class);
+        $mock->expects($this->any())
+            ->method('isInteractive')
+            ->will($this->returnValue($interactive));
+
+        if ($stream) {
+            $mock->expects($this->any())
+                ->method('getStream')
+                ->willReturn($stream);
+        }
+
+        return $mock;
+    }
+}
